requirements.txt
absl-py==1.4.0
accelerate==0.22.0
albumentations==1.1.0
attrs==23.1.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
contourpy==1.1.0
controlnet-aux==0.0.3
cycler==0.11.0
diffusers==0.20.0
dlib==19.24.1
einops==0.6.1
face-recognition==1.3.0
face-recognition-models==0.3.0
filelock==3.12.2
flatbuffers==23.5.26
fonttools==4.42.1
fsspec==2023.6.0
huggingface-hub==0.16.4
idna==3.4
imageio==2.31.1
importlib-metadata==6.8.0
importlib-resources==6.0.1
invisible-watermark==0.2.0
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
lazy_loader==0.3
MarkupSafe==2.1.3
matplotlib==3.7.2
mediapipe==0.10.0
mpmath==1.3.0
networkx==3.1
numpy==1.25.2
opencv-contrib-python==4.5.5.64
opencv-python==4.5.5.64
opencv-python-headless==4.5.5.64
packaging==23.1
Pillow==10.0.0
protobuf==3.20.3
psutil==5.9.5
pycparser==2.21
pyparsing==3.0.9
python-dateutil==2.8.2
PyWavelets==1.4.1
PyYAML==6.0.1
qudida==0.0.4
regex==2023.8.8
requests==2.31.0
safetensors==0.3.3
scikit-image==0.21.0
scikit-learn==1.3.0
scipy==1.11.2
six==1.16.0
sounddevice==0.4.6
sympy==1.12
threadpoolctl==3.2.0
tifffile==2023.8.25
timm==0.9.5
tokenizers==0.13.3
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.1
transformers==4.32.0
typing_extensions==4.7.1
urllib3==2.0.4
zipp==3.16.2
